[AIRFLOW-6933] Pass in env vars for all operators (#7576)
